backend/requirements.txt

fastapi
uvicorn
pydantic
python-dotenv
langdetect
chromadb
redis[async]
psycopg2-binary
asyncpg
httpx
